feat: network viz makes it easy to visualize if sync is complete
