+++ package.json
@@ -47,6 +47,7 @@
     "typescript": "^4.6.2"
   },
   "dependencies": {
+    "cli-table3": "^0.6.1",
     "ethers": "^5.6.1",
     "neverthrow": "^4.3.1"
   },
